+++ flutter/lib/illuminare.dart
@@ -5,5 +5,4 @@ export './src/outputs/outputs.dart';
 export './src/filters/filters.dart';
 
 export './src/illuminare.dart' show Illuminare;
-export './src/illuminare_logger.dart' show IlluminareLogger;
 export './src/illuminare_options.dart' show IlluminareOptions;