+++ flutter/lib/src/illuminare_log_event.dart
@@ -1,7 +1,9 @@
 import 'illuminare_trace.dart';
 
+/// The different log event levels
 enum IlluminareLogLevel { verbose, debug, info, warning, error, fatal }
 
+/// Log event contains all the information of a log event
 class IlluminareLogEvent {
   final IlluminareLogLevel level;
   final dynamic message;