+++ flutter/lib/src/illuminare_log_output.dart
@@ -1,5 +1,6 @@
 import 'illuminare_log_event.dart';
 
+/// Output log contains the level and lines of a log event
 class IlluminareOutputLog {
   final IlluminareLogLevel level;
   final List<String> lines;