+++ flutter/test/illuminare_test.dart
@@ -3,5 +3,9 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:illuminare/illuminare.dart';
 
 void main() {
-  test('test', () {});
+  test('Test illuminare can be initialized', () async {
+    await Illuminare.initialize();
+
+    expect(Illuminare.instance, isNotNull);
+  });
 }