+++ flutter/CHANGELOG.md
@@ -1,3 +1,10 @@
+# Changelog
+
 ## 0.0.1
 
 First version of Illuminare with working logging
+
+## 0.0.2
+
+Moved log methods to Illuminare class.
+Fixed minor warnings and did some cleanup